use populate to get student's classes
