+++ api/routes/class-routes.js
@@ -48,24 +48,6 @@ router.get('/classes/:id', async (req, res) => {
   }
 })
 
-// Get student's classes full details
-router.get('/students-classes/:id', async (req, res) => {
-  try {
-    const { id } = req.params;
-
-    if (!mongoose.Types.ObjectId.isValid(id)) {
-      return res.status(400).json({ error: 'Invalid ID' });
-    }
-
-    const classDetails = await User.findById(id)
-      .select('enrolledClasses')
-      .populate('enrolledClasses')
-    res.json(classDetails);
-  } catch (err) {
-    res.status(500).send(err);
-  }
-})
-
 // Create Class
 router.post('/classes', async (req, res) => {
   try {
